refactor(api): move settings to modules/ and share the project setting store
