--- apps/api/src/settings/hotkeys.ts
@@ -1,31 +1,0 @@
-import { t } from 'elysia';
-import { getSetting, setSetting } from '@repo/db';
-
-// The instance keyboard shortcuts (app_setting key 'hotkeys'): the combination
-// each command is bound to for everyone on this instance. Only the bindings
-// changed in god mode are stored; the web app fills the rest from its built-in
-// defaults, then applies the user's own overrides on top (user_preference.hotkeys).
-//
-// The set of commands lives in the web app (its lib/hotkeys), so the API stores the
-// map as given and only checks the shape: a command id and a combination written as
-// modifier tokens plus a key ('mod+k', 'n').
-
-const HOTKEYS_SETTING_KEY = 'hotkeys';
-
-export type HotkeyCombos = Record<string, string>;
-
-export const HotkeyCombosSchema = t.Record(
-  t.String({ pattern: '^[a-z][a-z0-9.-]{0,63}$' }),
-  t.String({ pattern: '^(mod\\+|shift\\+|alt\\+)*[a-z0-9]{1,10}$' }),
-);
-
-export async function getHotkeySettings(): Promise<HotkeyCombos> {
-  return (await getSetting<HotkeyCombos>(HOTKEYS_SETTING_KEY)) ?? {};
-}
-
-// Replaces the stored map. The god screen sends the full set of overrides, so an
-// unbound command is one left out rather than one written as empty.
-export async function setHotkeySettings(combos: HotkeyCombos): Promise<HotkeyCombos> {
-  await setSetting(HOTKEYS_SETTING_KEY, combos);
-  return combos;
-}
