+++ apps/api/src/shared/project-settings.ts
